Conflicts:
	src/Basescape/CraftSoldiersState.cpp
	src/Battlescape/BattlescapeState.cpp
	src/Battlescape/BattlescapeState.h
	src/Battlescape/Camera.cpp
	src/Battlescape/UnitInfoState.cpp
	src/CMakeLists.txt
	src/Engine/Flc.cpp
	src/Engine/Font.cpp
	src/Engine/Game.cpp
	src/Engine/Options.inc.h
	src/Engine/Palette.cpp
	src/Engine/Screen.cpp
	src/Engine/Surface.cpp
	src/Geoscape/Globe.cpp
	src/Interface/ArrowButton.cpp
	src/Interface/ScrollBar.cpp
	src/Interface/TextList.cpp
	src/Menu/IntroState.cpp
	src/Menu/OptionsVideoState.cpp
	src/Menu/StartState.cpp

+++ src/Basescape/TransferItemsState.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2013 OpenXcom Developers.
+ * Copyright 2010-2014 OpenXcom Developers.
  *
  * This file is part of OpenXcom.
  *
@@ -44,7 +44,7 @@
 #include "../Menu/ErrorMessageState.h"
 #include "TransferConfirmState.h"
 #include "../Engine/Options.h"
-#include "../aresame.h"
+#include "../fmath.h"
 
 namespace OpenXcom
 {
@@ -774,7 +774,7 @@ enum TransferType TransferItemsState::getType(size_t selected) const
  * @param selected Currently selected item.
  * @return Index of the selected item.
  */
-int TransferItemsState::getItemIndex(size_t selected) const
+size_t TransferItemsState::getItemIndex(size_t selected) const
 {
 	return selected - _soldiers.size() - _crafts.size() - _hasSci - _hasEng;
 }
